Data.BinP.PosP

Documentation

newtype PosP (b :: BinP)

PosP is to BinP is what Fin is to Nat, when n is Z.

Constructors

PosP

Fields

unPosP :: PosP' 'Z b

data PosP' (n :: Nat) (b :: BinP) where

PosP' is a structure inside PosP.

Constructors

AtEnd

Fields

:: Wrd n

-> PosP' n 'BE

position is either at the last digit;

Here

Fields

:: Wrd n

-> PosP' n ('B1 b)

somewhere here

There1

Fields

:: PosP' ('S n) b

-> PosP' n ('B1 b)

or there, if the bit is one;

There0

Fields

:: PosP' ('S n) b

-> PosP' n ('B0 b)

or only there if it is none.

Top & Pop

top :: SBinPI b => PosP b

top and pop serve as FZ and FS, with types specified so type-inference works backwards from the result.

>>> top :: PosP BinP4
0

>>> pop (pop top) :: PosP BinP4
2

>>> pop (pop top) :: PosP BinP9
2

pop :: (SBinPI a, Pred b ~ 'BP a, Succ a ~ b) => PosP a -> PosP b

See top.

Showing

explicitShow :: PosP b -> String

explicitShow' :: PosP' n b -> String

explicitShowsPrec :: Int -> PosP b -> ShowS

explicitShowsPrec' :: Int -> PosP' n b -> ShowS

Conversions

toNatural :: PosP b -> Natural

Convert PosP to Natural.

toNatural' :: forall n b. SNatI n => PosP' n b -> Natural

Convert PosP' to Natural.

Interesting

boring :: PosP 'BE

Counting to one is boring

>>> boring
0

Weakening (succ)

weakenRight1 :: SBinPI b => PosP b -> PosP (Succ b)

weakenRight1' :: forall b n. SBinP b -> PosP' n b -> PosP' n (Succ b)

Universe

universe :: forall b. SBinPI b => [PosP b]

>>> universe :: [PosP BinP9]
[0,1,2,3,4,5,6,7,8]

universe' :: forall b n. (SNatI n, SBinPI b) => [PosP' n b]

This gives a hint, what the n parameter means in PosP'.

>>> universe' :: [PosP' N.Nat2 BinP2]
[0,1,2,3,4,5,6,7]