Data.Wrd

Description

Fixed-Wrdth (unsigned) integers.

Documentation

data Wrd (n :: Nat) where

Fixed-width unsigned integers, Wrds for short.

The number is thought to be stored in big-endian format, i.e. most-significant bit first. (as in binary literals).

Constructors

WE :: Wrd 'Z

W0 :: Wrd n -> Wrd ('S n)

W1 :: Wrd n -> Wrd ('S n)

Showing

explicitShow :: Wrd n -> String

show displaying a structure of Wrd n

>>> explicitShow WE
"WE"

>>> explicitShow $ W0 WE
"W0 WE"

>>> explicitShow $ W1 $ W0 $ W1 $ W0 WE
"W1 $ W0 $ W1 $ W0 WE"

explicitShowsPrec :: Int -> Wrd n -> ShowS

showsPrec displaying a structure of Wrd n.

>>> explicitShowsPrec 0 (W0 WE) ""
"W0 WE"

>>> explicitShowsPrec 1 (W0 WE) ""
"(W0 WE)"

Conversions

toNatural :: Wrd n -> Natural

Convert to Natural number

>>> let u = W0 $ W1 $ W1 $ W1 $ W0 $ W1 $ W0 WE
>>> u
0b0111010

>>> toNatural u
58

>>> map toNatural (universe :: [Wrd N.Nat3])
[0,1,2,3,4,5,6,7]

Universe

universe :: forall n. SNatI n => [Wrd n]

All values, i.e. universe of Wrd .

>>> universe :: [Wrd 'Z]
[WE]

>>> universe :: [Wrd N.Nat3]
[0b000,0b001,0b010,0b011,0b100,0b101,0b110,0b111]

Bits

We have implementation of some Bits members, which doesn't need SNatI constraint.

xor :: Wrd n -> Wrd n -> Wrd n

(.&.) :: Wrd n -> Wrd n -> Wrd n

(.|.) :: Wrd n -> Wrd n -> Wrd n

complement :: Wrd n -> Wrd n

complement2 :: Wrd n -> Wrd n

complement2 w = complement w + 1

shiftR :: Wrd n -> Int -> Wrd n

shiftL :: Wrd n -> Int -> Wrd n

rotateL :: Wrd n -> Int -> Wrd n

rotateR :: Wrd n -> Int -> Wrd n

popCount :: Wrd n -> Int

setBit :: Wrd n -> Int -> Wrd n

clearBit :: Wrd n -> Int -> Wrd n

complementBit :: Wrd n -> Int -> Wrd n

testBit :: Wrd n -> Int -> Bool

Extras

shiftL1 :: Wrd n -> Wrd n

shiftR1 :: Wrd n -> Wrd n

rotateL1 :: Wrd n -> Wrd n

rotateR1 :: Wrd n -> Wrd n